Fixed order confirmation not recognized
Sometimes orders don't get recognized during confimation because the TickEvent is triggered  although the chest gui is not fully loaded yet.
With this fix the lastScreenDisplayName is only set if the gui is fully loaded and itemslot 13 is not null. By that a new order will only be initiated when the gui is fully loaded an all data can be gathered.